backend/requirements.txt
# FastAPI & Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database & ODM
motor==3.3.2
beanie==1.24.0
pymongo==4.6.1

# Authentication & Security
python-jose[cryptography]==3.3.0
bcrypt==4.1.2
python-dotenv==1.0.0
pydantic[email]==2.5.3
pydantic-settings==2.1.0

# Payment Gateway
razorpay==1.4.1

# Email
aiosmtplib==3.0.1
jinja2==3.1.3

# File handling & uploads
python-magic==0.4.27
Pillow==10.2.0

# Utilities
httpx==0.26.0
pydantic-extra-types==2.4.1
